stream read write is optimized, test app is optimized

+++ src/Message.h
@@ -17,20 +17,20 @@ private:
 	MessageHeader header{};
     MessageData data{};
 
-    bool readComponentList(int, std::vector<ComponentObject>&, MessageBlockHeader&, uint32_t&);
-    bool readJobName(int, char*, MessageBlockHeader&, uint32_t&);
-    bool readProcess(int, ProcessItem*, MessageBlockHeader&, uint32_t&);
-    bool readProcessID(int, long&, MessageBlockHeader&, uint32_t&);
-    bool readFile(int, ProcessFile&, MessageBlockHeader&, uint32_t&);
-    bool readMessageBlock(int in, MessageBlockHeader&, uint32_t&) override;
+    bool readComponentList(long, std::vector<ComponentObject>&, MessageBlockHeader&, uint32_t&);
+    bool readJobName(long, char*, MessageBlockHeader&, uint32_t&);
+    bool readProcess(long, ProcessItem*, MessageBlockHeader&, uint32_t&);
+    bool readProcessID(long, long&, MessageBlockHeader&, uint32_t&);
+    bool readFile(long, ProcessFile&, MessageBlockHeader&, uint32_t&);
+    bool readMessageBlock(long in, MessageBlockHeader&, uint32_t&) override;
 
-    bool writeComponentList(int, std::vector<ComponentObject>&, uint32_t&);
-    bool writeJobName(int, const char*, uint32_t&);
-    bool writeProcess(int, ProcessItem*, uint32_t&);
-    bool writeProcessID(int, long, uint32_t&);
-    bool writeFile(int, ProcessFile&, bool, uint32_t&);
+    bool writeComponentList(long, std::vector<ComponentObject>&, uint32_t&);
+    bool writeJobName(long, const char*, uint32_t&);
+    bool writeProcess(long, ProcessItem*, uint32_t&);
+    bool writeProcessID(long, long, uint32_t&);
+    bool writeFile(long, ProcessFile&, bool, uint32_t&);
 
-    bool writeMessageStream(int out, uint32_t&) override;
+    bool writeMessageStream(long out, uint32_t&) override;
 
     bool deSerializeHeader(const uint8_t*) override;
     bool serializeHeader(uint8_t *) override;
